ListResources always returns a string, sometimes base64 encoded.
Previously ListResources returned a string if no compression was
requested, and a base64 type object if compression was requested. That
is confusing, and probably relies on the fact that XML RPC offers a
base64 data type. Other comms mechanisms might not.

ListResources now returns a string in all cases. This string is
sometimes a base64 encoded, compressed rspec.