Added slice_name_to_urn to Framework_Base.
URNs may be site dependent, so the framework should be responsible for
creating the URN from a name. In the case of ProtoGENI, the URN
depends on the slice authority, and the user should not need to know
the syntax of a slice URN. The base class implementation just converts
directly to URN via namespace.long_urn. In ProtoGENI the domain of the
slice authority is computed based on the slice authority URL, and the
URN is constructed accordingly.