README-omni - more explanations - describe extending it
README
- more explanations
- document changes to init-ca (can generate each cert separately,
constants allow changing how they are generated)
- describe directory of federated certs to do federation
- document changes to calling gch (-r wants a directory of trusted
certs)
- document similar changes to calling gam
- mention Omni
- FIXME: Need refs to Wiki pages etc

src/client.py
- cleanup/ comments / documentation
- error check missing arguments
- better logger usage

src/gam.py
- cleanup / comments / documentation
- error check missing arguments
- better logger usage
- take the trusted CAs directory and produce a file for Python SSL to
use. It wants a single file, where SFA wants a directory of files.

src/gch.py
- cleanup / comments / documentation
- error check missing arguments
- better logger usage

src/geni/SecureXMLRPCServer
- comments

src/geni/__init__
- export CredentialVerifier

src/geni/am/am.py
- cleanup / comments / documentation
- error check missing arguments

src/geni/am/reference.py
- cleanup / comments / documentation
- error check missing arguments
- better logger usage
- pull out constants to top of file
- new CredentialVerifier.getCAsFileFromDir class method to turn a dir
of trusted certs into the single file that Python SSL expects
- more logging on verification failure and more informative error
messages to send back to the client
- logging at end of long AM methods to indicate success/completion
- Shutdown checks certificate validity - though there are no
privileges yet

src/geni/ch.py
- cleanup / comments / documentation
- error check missing arguments
- better logger usage
- pull out constants to top of file
- Aggregate Managers can be added to the CH via a new runtime method
call: register_aggregate_pair (not yet an XMLRPC method).
Currently the only calls to this are from the init() method
though
- runserver does the call to getCAsFileFromDir
- By default, CreateSlice no longer uses Josh's hack of using a single
commandline user cert. Instead, go through all of the trusted certs
given at startup, and test to see which if any signed the cert
of the XMLRPC client - the user. In this way, get the full cert chain
for that user and use that. This allows EG a PL user cert (if the PL
CH cert is trusted) to get a Slice credential from a GENI CH. Note
that the CA cert isn't enough - the intermediate CH cert is
needed. Hence the desire for CH certs to be trusted by each other.
If no signing cert is trusted, fall back on using the commandline user
cert. But note that this only works if the user who is the client is
the same as the user from the commandline. And if that's not a GCF
user then the AM will fail verify_issuer on the generated Slice
credential (what is a GCF CH doing signing a non GCF user?). So this
doesn't add much.

src/geni/omni/frameworks/framework_gcf.py
- error check arguments
- cache the user_cred return from get_user_cred in hopes of reducing
prompts for the user's password
- catch exceptions calling CH.CreateUserCredential so we can print a
more informative error message. As this is usually the first contact
with the CH, it's useful to know here what the error really is

src/geni/omni/frameworks/framework_pg
- error check arguments
- cache the user_cred return from get_user_cred in hopes of reducing
prompts for the user's password
- catch exceptions calling sa.GetCredential so we can print a
more informative error message. As this is usually the first contact
with the CH, it's useful to know here what the error really is

src/geni/omni/frameworks/framework_sfa
- error check arguments
- use logger
- cache the user_cred return from get_user_cred in hopes of reducing
prompts for the user's password
- catch exceptions calling registry.get_self_credential so we can print a
more informative error message. As this is usually the first contact
with the CH, it's useful to know here what the error really is

src/geni/omni/util/namespace
- error check args

src/init-ca
- comments / cleanup
- pull out constants to the top. Hopefully this makes it easier to
generate different certs with the same code - EG at Stanford
- create_cert parametrizes the subject and type of the cert better,
again for parametrization. Subject will be prefix+type+subj
- break up the make_certs method into methods for each of CA, CH, AM,
User allowing for a single class invocation to create only some certs,
or to be invoked repetitively for multiple AMs or Users. If only some
certs are being created, other signing certs are read back from files
if possible.
- parametrize the username in user certs
To support these, new options:
  -u username
  -notAll means dont genenerate all certs. Use with other args:
  --ca means do the CA, etc for --ch --am and --exp for experimenter
certs

src/omni.py
- documentation & cleanup
- error check args
- more/better use of logger
- listresources handling of geni_compressed flag was
inconsistent/broken. The result of listresources needs to be
decompressed if the flag was set. EXCEPT that in the PG framework
apparently the resource list is not compressed even when it was
requested, so catch that error

src/sfa/trust/certificate
- add a log call, and tweak an exception text

src/sfa/trust/credential
- trim whitespace
- add more useful info to the verify CredentialNotVerifiable
exceptions
- comments
- credential.dump bug: was calling self.get_subject() which isnt there

src/sfa/trust/gid
- turn down a log statement
- add another log statement
- bug: call to GidParentHrn should be RAISING that
exception. Otherwise we were silently allowing a cert that should be an
error