Added status to resources and slivers.
Resources now have a status field that contains one of five
states. Resource status is set in CreateSliver (ready) and
DeleteSliver (unknown). Sliver status is determined by examining the
status of all the resources. If any resource is shutdown, failed, or
configuring, then that is the status of the slice (in that order). If
all resources are ready, the sliver is ready. Otherwise the sliver is
status is unknown.