Added -n and -a args to omni.
The -a argument allows a user to specify a single aggregate on the
command line. If specified, omni uses only that aggregate and ignores
the aggregates in the config file and via the Clearinghouse.

The -a argument affects all calls.

The -n argument is for "native" mode with RSpecs. It requires the -a
argument. If -n is specified the RSpec is not converted to an
omnispec.

Only ListResources handles the -n argument at this time.