Fix Python 2.6 compatibility.
Python 2.7 introduces an incompatible change in the contract of
xmlrpc.Transport.make_connection. In Python 2.6, make_connection
returns an httplib.HTTP instance. In Python 2.7, make_connection
returns an httplib.HTTPConnection instance. SafeTransportWithCert now
looks at the Python version number to decide which kind of instance to
return. It's ugly, but appears to work.