Fix download of sfa GID.
This was broken with the M2Crypto transition. There is a race
condition in the case where the GID has not yet been downloaded. The
Framework_Base constructor is called late, which means the sslctx
attribute has not yet been initialized. The base class constructor is
now called after the self-signed cert is created so that the base
class can be initialized. The base class is initialized again after
the GID is downloaded to clear out the old SSL Context which uses the
temporary self-signed cert instead of the real GID.