Back out M2Crypto.
Move M2Crypto off to the side and return to the Python SSL library for
SSL support. The M2Crypto implementation is problematic in different
ways in different environments. While the single password feature is
desirable, M2Crypto appears to introduce low-level issues that are
hard to resolve and vary by platform/Python version/M2Crypto version.

This commit also introduces a layer in the framework classes to funnel
all XML-RPC client creation through a single method in the base
class. This should make a future transition a little bit easier.