Issue #1: ProtoGENI puts namespace attributes on the signed-credential root element. When producing a delegated credential, those attributes must be preserved, or else the parent (ProtoGENI) credential will cease to verify - xmlsec1 says the data and digest do not match.
Solution: encode() now copies any attributes on the parent credential signed-credential element to the new credential.

Issue #2:
ProtoGENI uses '*' for its only privilege.
decode() has special logic to handle this.
There is currently a bug though, where that special logic fails to copy the 'delegatable' flag to all the implied rights - effectively making them look non-delegatable, regardless of what the original credential said.
The effect is that a PG credential delegated to a PLC user looks invalid, because this code thinks the parent credential was not delegatable.

Solution: Inferred rights have their delegate bit set per the original single '*' privilege's value.

Non-issues but places I have made changes:
1) Useful for my debugging: Improve error messages on verify failures
  a) xmlsec1 message includes the refID that failed to validate
  b) ChildRightsNotSubsetOfParent includes the refIDs and text to explain

2) Comments indicating differences with ProtoGENI

3) Fix whitespace / formatting on the Signature template to clean it up a little. Clearly not important

4) Useful for my delegation script: Allow verify to get a None for trusted_certs and not have errors
This is strange, but I had a reason.
Obviously you usually wouldn't do this.
But I wanted to re-use the verify() method in my delegateSliceCred.py script.
And I wanted to allow the user to not have the right root certificates available (that verify() would want), and still get the benefit of the other checks that verify() does for you.
So this change has a bunch of 'if trusted_certs is not None'