Make abac an optional import.
If the import of ABAC or Creddy fail, catch the error and
continue. For almost all users, that's acceptable. It will only be
confusing to the very small handful of users who are experimenting
with ABAC. They'll see an error soon enough that should indicate that
there's a problem loading ABAC, and they'll know how to handle it.