Add FIXMEs about changing api version.
There are numerous places where omni changes the declared api_version
if one of the aggregates speaks a different version. This could cause
problems if there are multpiple aggregates and one of the early ones
has a version mismatch. Note that at each place where we change the
api_version option (effectively a global) to that it can be fixed
later.