Do not modify ssh config file.
omni-configure.py script by default adds an
IdentityFile line for the private SSH keys so that
they are included when the user tries to ssh to hosts.
This is probably not very useful and there is no flag
to turn this off. So it is commented out for now,
if we later need this again we can see what is the best
way to do it.