keep resources' states updated in am2
This is my suggested fix for issue #848.

In CreateSliver() the resources' *available* attribute was never set
to False, allowing users to 'overbook' am's resources.

In DeleteSliver() there was an attempt to reset resources to their
default state, but the loop was never executed due to the list always
being empty. The list of resources in a Sliver is cleared when
deallocating the aggregate's *slice_urn* container because the variable
*resources* is a mere reference shared with the aggregate's, instead
of a full copy.
To fix this, loop over the resources before deallocating the list, and
since there's a method called reset(), use it to make the purpouse of
the loop obvious instead of manually resetting the resources to the
default state.