Update key size and signature algorithm.
Change default key size from 1024 to 2048. Change default signature
algorithm from md5 to sha256. CentOS 7 refuses SSL connections when
the server uses a certificate with an md5 signature
algorithm. Updating to sha256 fixes the issue. While we're there,
update to larger keys as well since 1024 is on the way out.