Issue #873: Fix combining stitching path using ExoGENI AMs
ExoGENI manifests list different stitching hops than we request (or that other AMs will remember in their manifests). In addition, ExoGENI's initial manifest does not list the VLAN tag that you were given (not until the reservation is no longer configuring).
So add more special case code to detect these situations, and edit the template stitching hops to use the proper vlan tags on ExoGENI hops. We do not try to list the stitching path hop that the ExoGENI AM lists, as typically that is a single hop and we previously had 3, so the hop #s wouldn't match and that would be a larger edit.

	modified:   CHANGES
	modified:   src/gcf/omnilib/stitch/ManifestRSpecCombiner.py