Delegate TLS handling to a frontend
A per python docs, the built-in HTTP request handler is not fit for
production. Add an option to use geni-tools in production by having it
listen on localhost, using http, to be called by a frontend that will
handle the TLS handshake and pass the client certificate as an HTTP
Header.

To use this variant, you need to set proto and certheader options in the
configuration file.